+++ lib/date_time.dart
@@ -1,7 +1,7 @@
 library date_time;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export './res/date.dart';
+export './res/date_range.dart';
+export './res/extensions.dart';
+export './res/time.dart';
+export './res/time_range.dart';