+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* [Add] `Date` and `DateRange` classes
+* [Add] `Time` and `TimeRange` classes
+* [Add] extensions for `DateTime` to separate `Date` & `Time`